requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
djangorestframework==3.15.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
